requirements.txt
Flask==2.2.5
mysql-connector-python==8.0.33
Werkzeug==2.2.3
Jinja2==3.1.6
click==8.1.8
itsdangerous==2.1.2
gunicorn==20.1.0
